Refactor: Move our game functions out of the Game component and into context
This will help to keep the Game component tidier as we can now
break it up into smaller components and share the state without having
to prop drill. As prop drilling meant all of the functions would need to sit
in the Game component, making it harder to parse
