+++ lib/game/GameContext.js
@@ -1,0 +1,267 @@
+import { createContext, useState, useEffect } from 'react'
+import { supabase } from '../../utils/supabaseClient'
+import { useAuth } from '../auth/useAuth'
+import useActiveGame from './useActiveGame'
+import usePlayersInGame from './usePlayersInGame'
+import usePin from './usePin'
+import createPinNumber from '../../utils/createPinNumber'
+
+const GameContext = createContext(null)
+
+const GameProvider = ({ children }) => {
+  const [loading, setLoading] = useState(false)
+  const [activeGame, setActiveGame] = useActiveGame()
+  const [players, setPlayers] = usePlayersInGame()
+  const [pin, setPin] = usePin()
+  const { user } = useAuth()
+
+  /**
+   * Check our pin number against all of the currently live games.
+   * If the pin exists elsewhere regenerate it so we don't ever have two
+   * games with the same access pin
+   * @returns
+   */
+  const createGamePin = async () => {
+    try {
+      const pinNumber = createPinNumber()
+
+      // If there is a matching pin recursively run this function again
+      const { data, error, status } = await supabase
+        .from('games')
+        .select('in_progress, pin')
+        .match({
+          in_progress: true,
+          pin: pinNumber,
+        })
+
+      if (error) console.error(status, error)
+
+      if (data.length > 0) {
+        createGamePin(pinNumber)
+      }
+
+      // Return new pin number
+      return pinNumber
+
+      // Loop over result
+    } catch (error) {
+      console.error(error)
+    }
+  }
+
+  /**
+   * Creates a new game instance
+   *
+   * Sets the pin number and the active game to our state
+   */
+  const createGame = async () => {
+    try {
+      setLoading(true)
+
+      const gamePin = await createGamePin()
+
+      const { data, error, status } = await supabase.from('games').insert({
+        creator_id: user.id,
+        pin: gamePin,
+        in_progress: true,
+        players_in_game: JSON.stringify([{ id: user.id }]),
+      })
+
+      if (error) console.error(status, error)
+
+      setPin(gamePin)
+      setActiveGame(gamePin)
+    } catch (error) {
+      console.error(error)
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  // TODO: Would be worth deleting the table row once the game is over so we don’t end up with a massive db to loop over
+  /**
+   * Ends the current game
+   *
+   * Clears the state of our pin number, the active game and the players array
+   *
+   * @param {number} pin active game pin number
+   */
+  const endGame = async (pin) => {
+    try {
+      setLoading(true)
+
+      const { error, status } = await supabase
+        .from('games')
+        .update({
+          in_progress: false,
+          players_in_game: null,
+        })
+        .match({ creator_id: user.id, pin })
+
+      if (error) console.error(status, error)
+
+      // Remove our saved pin number
+      setPin('')
+      setActiveGame('')
+      setPlayers([])
+      // TODO: Remove subscription
+    } catch (error) {
+      console.error(error)
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  /**
+   * Adds player to the current game session
+   *
+   * @param {Event} e
+   */
+  const joinGame = async (e) => {
+    // TODO: When joins successfully redirect player to the game screen
+    e.preventDefault()
+    // Join pin input values
+    let pinInputValue = ''
+
+    const pinInputs = [...e.target.elements].filter(
+      (input) => !input.matches('[type="submit"]'),
+    )
+
+    for (const input of pinInputs) {
+      pinInputValue += input.value
+    }
+
+    try {
+      setLoading(true)
+
+      // Check to make sure our player isn't already in the game
+      const activePlayers = await supabase
+        .from('games')
+        .select('players_in_game')
+        .match({ in_progress: true, pin: pinInputValue })
+
+      // Turn our string of players into an array
+      const { players_in_game } = activePlayers.data[0]
+      const playersJSON = JSON.parse(players_in_game)
+
+      for (const player of playersJSON) {
+        if (player === '') continue // Skip any blank strings
+
+        // if user id doesn't match player id add them to the players list
+        if (user.id !== player.id) {
+          playersJSON.push({ id: user.id })
+        }
+      }
+
+      const updatePlayers = await supabase
+        .from('games')
+        .update({
+          players_in_game: JSON.stringify(playersJSON),
+        })
+        .match({ in_progress: true, pin: pinInputValue })
+
+      if (updatePlayers.error)
+        console.error(updatePlayers.status, updatePlayers.error)
+
+      setActiveGame(pinInputValue)
+    } catch (error) {
+      console.error(error)
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  /**
+   * Get the players currently in the game when the activeGame state updates and
+   * push the returned player array into our players state.
+   */
+  useEffect(() => {
+    let mounted = true
+
+    if (mounted) {
+      // TODO: A lot of boilerplate code going on -- could I refactor?
+      const getPlayersInGame = async () => {
+        try {
+          const {
+            data: games,
+            error,
+            status,
+          } = await supabase.from('games').select('players_in_game').match({
+            in_progress: true,
+            pin: activeGame,
+          })
+
+          if (error) console.error(status, error)
+
+          if (games.length > 0) {
+            // Turn our string of players into an array
+            const playersArr = JSON.parse(games[0].players_in_game)
+
+            setPlayers(playersArr)
+          }
+        } catch (error) {
+          console.error(error)
+        }
+      }
+
+      getPlayersInGame()
+    }
+
+    // Async side effect cleanup
+    return () => (mounted = false)
+  }, [activeGame, setPlayers])
+
+  /**
+   * Listen to the row in our supabase table for updates and pass the returned payload
+   * into our handleNewPlayer function.
+   *
+   * We're listening to the row in the games table that has a pin equal to our current
+   * pin number.
+   */
+  useEffect(() => {
+    const liveGameSubscription = supabase
+      .from(`games:pin=eq.${pin}`)
+      .on('*', (payload) => {
+        console.log('Change received!')
+        handleNewPlayer(payload)
+      })
+      .subscribe()
+
+    return () => supabase.removeSubscription(liveGameSubscription)
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [pin, players])
+
+  /**
+   * Push players who join the game into the players array
+   *
+   * @param {object} payload
+   */
+  const handleNewPlayer = (payload) => {
+    console.log({ payload }) // TODO: Handle errors?
+
+    const { players_in_game } = payload.new
+
+    if (players_in_game) {
+      const playersJSON = JSON.parse(players_in_game)
+      setPlayers(playersJSON)
+    }
+  }
+
+  return (
+    <GameContext.Provider
+      value={{
+        createGame,
+        endGame,
+        joinGame,
+        activeGame,
+        pin,
+        players,
+        loading,
+      }}
+    >
+      {children}
+    </GameContext.Provider>
+  )
+}
+
+export { GameContext, GameProvider }
